Keep GH action scheduled builds alive
GitHub will suspend the scheduled trigger for GitHub action workflows if
there is no commit in the repository for the past 60 days. The cron based
triggers won't run unless a new commit is made.  We use the Keepalive
Workflow[1] action to commit to make new commits to the repo so that
GH actions won't suspend the builds.

[1] https://github.com/gautamkrishnar/keepalive-workflow